requirements.txt
Pillow==2.2.1
ShopifyAPI==1.0.7
requests==2.0.1
